Refactor import statements to use absolute paths for tracing configuration across multiple modules
