+++ backend/core/ASR/src/asr_infrence.py
@@ -1,7 +1,7 @@
 import torch
 from backend.core.ASR.src.preprocess_audio import audio_utils
 from backend.core.ASR.src.load_model import LoadSeamlessModel
-from backend.core.ASR.src.tracing_config import trace_cleanup_operation
+from backend.core.tracing_config import trace_cleanup_operation
 from dotenv import load_dotenv
 import os
 from langsmith import traceable
