+++ backend/core/ASR/src/pipeline.py
@@ -6,7 +6,7 @@ from typing import List
 from backend.core.ASR.src.asr_infrence import transcribe
 from backend.core.ASR.src.llm_engine import LLMEngine
 from backend.core.ASR.src.models import PipelineOutput, TranscriptionSegment
-from backend.core.ASR.src.tracing_config import setup_tracing, get_trace_metadata, is_tracing_enabled, trace_file_operation, trace_cleanup_operation
+from backend.core.tracing_config import setup_tracing, get_trace_metadata, is_tracing_enabled, trace_file_operation, trace_cleanup_operation
 from langsmith import traceable
 
 # Configure logging
